requirements.txt
fastapi
uvicorn
groq
pandas
fpdf2
openpyxl
python-docx
python-multipart
python-dotenv
aiofiles
